Refactor User ID to use Discord Snowflake ID as Primary Key
- Changed `User.id` from Auto-Increment Integer to BigInteger.
- Removed `discord_id` column from `User` model and database schema.
- Updated foreign keys in `Planet`, `Fleet`, `Wallet`, `Contract` models to `BigInteger`.
- Updated `init.sql` schema definition.
- Updated `AuthService` to use Discord ID as the primary key.
- Updated `read_users_me` endpoint to remove `discord_id` field.
- Updated tests to use random BigInteger IDs.

+++ backend/app/models/contract.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime, BigInteger
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.models.base import Base
@@ -7,8 +7,8 @@ class Contract(Base):
     __tablename__ = "contracts"
 
     id = Column(Integer, primary_key=True, index=True)
-    issuer_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    contractor_id = Column(Integer, ForeignKey("users.id"), nullable=True)
+    issuer_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    contractor_id = Column(BigInteger, ForeignKey("users.id"), nullable=True)
 
     origin_planet_id = Column(Integer, ForeignKey("planets.id"), nullable=False)
     destination_planet_id = Column(Integer, ForeignKey("planets.id"), nullable=False)
