+++ backend/app/models/wallet.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey, DateTime, String, UniqueConstraint
+from sqlalchemy import Column, Integer, Float, ForeignKey, DateTime, String, UniqueConstraint, BigInteger
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.models.base import Base
@@ -7,7 +7,7 @@ class Wallet(Base):
     __tablename__ = "wallets"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"), unique=True, nullable=False)
+    user_id = Column(BigInteger, ForeignKey("users.id"), unique=True, nullable=False)
 
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
